requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-crispy-forms==1.8.1
gunicorn==20.0.2
idna==2.8
Pillow==6.2.1
pkg-resources==0.0.0
pydub==0.23.1
pytz==2019.2
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.6
